Bug fix concering live feed
I tried to fix the issue raised here
(https://community.backtrader.com/topic/623/anyone-use-backtrader-to-do-live-trading-on-bitcoin-exchange/199).
Mainly, the code got the ohlcv data of the current period, before it
had completely finished, meaning trades where missing. Currently the
code assumes that the timestamp is from the beginning of the
candleperiod, which is not always the case, but this is a ccxt issue and
I raised it here (https://github.com/ccxt/ccxt/issues/2206).